backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic[email]>=2.4.2
pydantic-settings>=2.0.3
scikit-learn>=1.3.0
pandas>=2.1.0
joblib>=1.3.0
google-generativeai>=0.3.0
openai>=1.0.0
spacy>=3.7.0
